+++ Tests/SentryTests/Integrations/Performance/Network/SentryNetworkTrackerIntegrationTests.swift
@@ -6,6 +6,7 @@ class SentryNetworkTrackerIntegrationTests: XCTestCase {
     
     private static let dsnAsString = TestConstants.dsnAsString(username: "SentryNetworkTrackerIntegrationTests")
     private static let testURL = URL(string: "http://localhost:8080/hello")!
+    private static let testTraceURL = URL(string: "http://localhost:8080/echo-sentry-trace")!
     private static let transactionName = "TestTransaction"
     private static let transactionOperation = "Test"
     
@@ -111,7 +112,7 @@ class SentryNetworkTrackerIntegrationTests: XCTestCase {
         let traceContext = transaction.traceContext
         
         if canHeaderBeAdded() {
-            let expected = [SENTRY_TRACE_HEADER: transaction.toTraceHeader().value(), SENTRY_BAGGAGE_HEADER: traceContext.toBaggage().toHTTPHeader() ]
+            let expected = [SENTRY_BAGGAGE_HEADER: traceContext.toBaggage().toHTTPHeader() ]
             XCTAssertEqual(expected, configuration.httpAdditionalHeaders as! [String: String])
         } else {
             XCTAssertNil(configuration.httpAdditionalHeaders)
@@ -135,7 +136,7 @@ class SentryNetworkTrackerIntegrationTests: XCTestCase {
         }
         
         if canHeaderBeAdded() {
-            let expected = [SENTRY_TRACE_HEADER: transaction.toTraceHeader().value(), SENTRY_BAGGAGE_HEADER: traceContext.toBaggage().toHTTPHeader()]
+            let expected = [SENTRY_BAGGAGE_HEADER: traceContext.toBaggage().toHTTPHeader()]
                 .merging(additionalHeaders) { (current, _) in current }
             XCTAssertEqual(expected, dataTask.currentRequest?.allHTTPHeaderFields)
         } else {
@@ -227,6 +228,28 @@ class SentryNetworkTrackerIntegrationTests: XCTestCase {
         XCTAssertEqual("GET \(SentryNetworkTrackerIntegrationTests.testURL)", networkSpan.context.spanDescription)
         
         XCTAssertEqual("200", networkSpan.tags["http.status_code"])
+    }
+    
+    func testGetRequest_CompareSentryTraceHeader() {
+        startSDK()
+        let transaction = SentrySDK.startTransaction(name: "Test Transaction", operation: "TEST", bindToScope: true) as! SentryTracer
+        let expect = expectation(description: "Request completed")
+        let session = URLSession(configuration: URLSessionConfiguration.default)
+        var response: String?
+        let dataTask = session.dataTask(with: SentryNetworkTrackerIntegrationTests.testTraceURL) { (data, _, _) in
+            response = String(data: data ?? Data(), encoding: .utf8) ?? ""
+            expect.fulfill()
+        }
+        
+        dataTask.resume()
+        wait(for: [expect], timeout: 5)
+        
+        let children = Dynamic(transaction).children as [SentrySpan]?
+        
+        XCTAssertEqual(children?.count, 1) //Span was created in task resume swizzle.
+        let networkSpan = children![0]
+        
+        XCTAssertEqual(networkSpan.toTraceHeader().value(), response)
     }
     
     private func asserrtNetworkTrackerDisabled(configureOptions: (Options) -> Void) {